底盘position下的陀螺仪 修bug

+++ User/Inc/RTOS.h
@@ -8,7 +8,7 @@
 #include "stm32f4xx_hal.h"
 #include "cmsis_os.h"
 
-extern uint32_t debug;
+extern float debug;
 
 /*队列*/
 extern osMessageQueueId_t CAN1SendQueueHandle;
