+++ User/Src/RTOS.c
@@ -4,4 +4,4 @@
 
 #include "RTOS.h"
 
-uint32_t debug;
+float debug;
